CS-1
I want to center the box so I need to get a hold of the dimensions of the visible page. So I added a javascript function to get the height and width and call that method inside our home page and use it to center the dialog box. I created 2 internal classes to hold the window dimensions and the styling for the dialog box container.